+++ artiste/src/vsx_widget/lib/vsx_widget_base_edit.cpp
@@ -143,7 +143,7 @@ void vsx_widget_base_edit::process_line(int n_line) {
   if (lines[i][0] == '/' && lines[i][1] == '/')
   fill = (char)0x02;
 
-  for (int j = 0; j < lines[i].size(); ++j) {
+  for (size_t j = 0; j < lines[i].size(); ++j) {
     b = b + fill;
   }
   if (fill == (char)0x01)
@@ -170,7 +170,7 @@ void vsx_widget_base_edit::process_lines() {
     if (lines[i][0] == '/' && lines[i][1] == '/')
     fill = (char)0x02;
 
-    for (int j = 0; j < lines[i].size(); ++j) {
+    for (size_t j = 0; j < lines[i].size(); ++j) {
       b = b + fill;
     }
     if (fill == (char)0x01)
@@ -187,7 +187,7 @@ void vsx_widget_base_edit::process_lines() {
 }
 
 void vsx_widget_base_edit::calculate_scroll_size() {
-  int t_longest_line = 0;
+  size_t t_longest_line = 0;
   for (std::vector<vsx_string>::iterator it = lines.begin(); it != lines.end(); ++it) {
     if ((*it).size() > t_longest_line) t_longest_line = (*it).size();
   }
@@ -346,7 +346,7 @@ void vsx_widget_base_edit::event_mouse_wheel(float y)
 
 int count_whitespaces(vsx_string& s)
 {
-	int i = 0;
+	size_t i = 0;
 	while (i < s.size() && s[i] == ' ')
 	{
 		//printf("line: %d\n",__LINE__);
@@ -664,7 +664,7 @@ bool vsx_widget_base_edit::event_key_down(signed long key, bool alt, bool ctrl, 
 #endif
 #endif
       ++caretx;
-      if (caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x) {
+      if ((size_t)caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x) {
 #ifdef _WIN32
         event_key_down(-40,false,false,false);
         event_key_down(-36,false,false,false);
@@ -697,7 +697,7 @@ bool vsx_widget_base_edit::event_key_down(signed long key, bool alt, bool ctrl, 
             //fix_pos();
           }
         }
-        if (caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x)
+        if ((size_t)caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x)
 #ifdef _WIN32
           event_key_down(-35);
 #else
@@ -743,7 +743,7 @@ bool vsx_widget_base_edit::event_key_down(signed long key, bool alt, bool ctrl, 
           ++scroll_y;
           --carety;
         }
-        if (caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x)
+        if ((size_t)caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x)
          caretx = lines[carety+(int)scroll_y].size()-(int)scroll_x;
       }
       break;
@@ -875,7 +875,7 @@ bool vsx_widget_base_edit::event_key_down(signed long key, bool alt, bool ctrl, 
         command_q_b.add_raw(d+lines[0]);
         parent->vsx_command_queue_b(this);
       } else {
-        if (caretx+(int)scroll_x > lines[carety+(int)scroll_y].size()) event_key_down(-35,false,false,false);
+        if ((size_t)caretx+(size_t)scroll_x > lines[carety+(int)scroll_y].size()) event_key_down(-35,false,false,false);
         while (c2 < carety+(int)scroll_y) { ++c2; ++it; ++itp; ++itlv; }
         ++it;
         ++itp;
@@ -931,7 +931,7 @@ bool vsx_widget_base_edit::event_key_down(signed long key, bool alt, bool ctrl, 
       	lines[carety+(int)scroll_y].insert(caretx+(int)scroll_x,(char)key);
       	updates++;
       	++caretx;
-      	if (caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x)
+      	if ((size_t)caretx > lines[carety+(int)scroll_y].size()-(int)scroll_x)
        	caretx = lines[carety+(int)scroll_y].size()-(int)scroll_x;
       	int t_scroll_x = (int)scroll_x;
       	if (caretx+(int)scroll_x > characters_width) ++scroll_x;
